ml_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
scikit-learn==1.3.2
xgboost==2.0.2
pandas==2.1.3
numpy==1.26.2
influxdb-client==1.38.0
python-dotenv==1.0.0
joblib==1.3.2
